04-stateful-workflows/pyproject.toml
[project]
name = "04-stateful-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "pandas>=2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]
